backend/requirements.txt
# Django and REST Framework
Django==5.1.5
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==24.3

# Database
psycopg2-binary==2.9.10

# Celery and Redis
celery==5.5.3
redis==6.4.0
django-celery-beat==2.8.1
django-celery-results==2.6.0

# Twilio for WhatsApp
twilio==9.8.3

# OpenAI (also used for OpenRouter API)
openai==2.1.0

# LangChain and Deep-Agent
deepagents==0.2.7
langgraph==1.0.3
langchain==1.0.7
langchain-core==1.0.5
langchain-openai==1.0.3
langchain-anthropic==1.0.4

# Environment variables
python-decouple==3.8

# Utilities
requests==2.32.5
python-dateutil==2.9.0.post0
Pillow==11.1.0

# Production server
gunicorn==23.0.0

# Development tools
black==25.9.0
flake8==7.1.1
mypy==1.18.2

# Type stubs
django-stubs==5.2.5
djangorestframework-stubs==3.16.4
celery-types==0.23.0
types-redis==4.6.0.20241004
types-requests==2.32.4.20250913
types-python-dateutil==2.9.0.20250822
